create modal and populate with movie data
I created the modal using React Portals, and I've
also finished coding the functionality for it.
When a movie is clicked, its data is loaded into
the modal. For the website to know WHICH movie's data to
load in, I first grabbed the ID of the clicked movie and
stored it in state.