implement movie search feature and responsive design
The user is now able to search for movies! I
used a dethrottle function to limit the amount of
API calls made while the user types. I also added
a few "default values" to be displayed if they're not
available, such as the movie description or
image.

I also made the website responsive.